+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Fixed
 - run_program:  Fixed problem with mutable default arguments issue.
 
+### Added
+- create_rq:  Function to create and return a RabbitMQ instance.
+
 
 ## [0.3.5] - 2019-03-01
 ### Added