+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Added
 - create_rq:  Function to create and return a RabbitMQ instance.
 
+### Changed
+- process_message:  Refactored function and replaced instance code with create_rq call.
+
 
 ## [0.3.5] - 2019-03-01
 ### Added