+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - run_program:  Fixed problem with mutable default arguments issue.
 
 ### Added
+- camelize:  Camelcases a string.
 - create_rq:  Function to create and return a RabbitMQ instance.
 
 ### Changed